pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "jobs-cli"
version = "0.1.0"
description = "CLI tool to find product leadership job listings from ATS platforms"
requires-python = ">=3.9"
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "pyyaml>=6.0",
    "googlesearch-python>=1.2",
    "requests>=2.28",
    "beautifulsoup4>=4.12",
]

[project.scripts]
pjobs = "jobs.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["jobs*"]
